Corrected conversion tfrecord_to_dat.py code: specifying dtype=np.float32 in embedding = np.array(example.features.feature['embedding'].float_list.value, dtype=np.float32)] was necessary to handle the embeddings properly, which corrected the shape discrepancies (previously, embeddings with a shape of (2752,) were generated, which is double the expected (1376,)).

+++ utils/converter_utils/tfrecord_to_dat.py
@@ -12,7 +12,7 @@ def parse_tfrecord_file(tfrecord_filepath):
     for raw_record in raw_dataset.take(1):   # Only take the first record
         example = tf.train.Example()
         example.ParseFromString(raw_record.numpy())
-        embedding = np.array(example.features.feature['embedding'].float_list.value)
+        embedding = np.array(example.features.feature['embedding'].float_list.value, dtype=np.float32)
         return embedding   # Return embedding after parsing the first record
